fix(observability): remove boot-crashing URL validation + DRY pgboss SQL
- ALERT_WEBHOOK_URL: drop .url() validation that crashed app boot on
  invalid webhook — the app never uses this value at runtime (Grafana
  container reads it directly from docker-compose env)
- Extract collectQueueDepths() into metrics.ts — /metrics and /readyz
  both queried pgboss.queue with identical SQL, now share one function
- Strip "what" comments from metrics.ts and app.ts per AGENTS.md
